back/requirements.txt
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
numpy==1.26.4
pillow==11.3.0
matplotlib==3.10.6
scikit-image==0.25.2
captum==0.8.0
streamlit==1.50.0
ipykernel==6.30.1
efficientnet-pytorch==0.7.1
lime==0.2.0.1
wandb==0.17.5
scikit-learn==1.4.2
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6